Fix typo in print statement (FORWARD example)
The sample code for showcasing FORWARD computation (in the K axis) calls
`mult_upward` with `origin=(nhalo,nhalo,1)`. The print statement differs
in the z-value and shows `origin=(nhalo,nhalo,0)`. This can be confusing
for developers which just skim the output and don't read the code that
is actually executed.