No threshold / perturbation analysis by default
Currently, when translate tests fail, we run 10 (by default) times the
same translate test/savepoint again with slightly perturbed inputs. The
outputs are then analyzed to give developers an indication on whether
errors that they are seeing are in the expected range of "noise
propagation" or if this is more likely to be an error in porting.