Translate tests: no threshold analysis by default (#134)
* No threshold / perturbation analysis by default

Currently, when translate tests fail, we run 10 (by default) times the
same translate test/savepoint again with slightly perturbed inputs. The
outputs are then analyzed to give developers an indication on whether
errors that they are seeing are in the expected range of "noise
propagation" or if this is more likely to be an error in porting.

* Rename {PACE -> NDSL}_TEST_N_THRESHOLD_SAMPLES
* Unrelated fix typos
- fix typos
- capitalize words at the start of a sentence
- add point a point at the end of a sentence

---------

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>