Run tests when pushing to protected branches
We don't use a merge queue and we allow to merge outdated branches for
practical reasons (because in particular pyFV3 and pace tests take a
long time to run). It could thus happen that we get a bad merge. In that
unlikely case, we'd like to know and thus run tests when pushing to
protected branches (`main` and `develop`) as a safety net.