refactor: remove workaround in fill_corners_dgrid
Due to a gt4py-dace bridge around argument lists and pruning unused
arguments/statements, we introduced non-sensical code in this stencil to
avoid the stencil being completely empty, which would trigger "missing"
arguments.

See PRs / issues

https://github.com/NOAA-GFDL/NDSL/pull/60
https://github.com/NOAA-GFDL/NDSL/pull/69
https://github.com/NOAA-GFDL/NDSL/issues/42
https://github.com/NOAA-GFDL/NDSL/issues/70

for context.

Since the re-write of the gt4py-dace bridge (based on "schedule trees"),
we don't have this limitation anymore and the workaround can be removed
again.