fix: boilerplate allocation (layout warning)
`ndsl/boilerplate.py` has two functions to simplify creation of
`StencilFactory` and `QuantityFactory`. Both factories depend on the
backend where the numpy backend acts a fallback for both. The backend
wasn't properly forwarded to the `QuantityFactory` leading to data that
wasn't necessarily optimially aligned (except e.g. for the numpy
backend). For example, in case of the `dace:cpu` backend, this leads to
performance warning messages about improperly aligned memory.

This is what Christopher Harrop was seeing because he specified the
`dace:cpu` backend, which wasn't properly forwarded to the
`QuantityFactory` leading to warning messages about improperly aligned
memory.