Breaking change: QuantityFactory from GridSizer and backend name
Change `QuantityFactory` to initialize from a `GridSizer` (as
previously) and a backend name (new). This effectively hides the
previous `numpy` argument, which is effectively an internal allocator
that users shouldn't need to know about. It's basically what
`from_backend()` was doing before (which is now obsolete and was thus
removed).

This is a BREAKING CHANGE and users will need to update their codes if
they instantiated QuantityFactories themselves. For users relying on the
`boilerplate` module, no changes need to happen.