BREAKING CHANGE: orchestration boilerplate options
This PR introduces a breaking change in the orchestration boilerplate
funtions that sets up `StencilFactory` and `QuantityFactory`. The option
`on_cpu` is replaced by a `backend` option. The backend option defaults
to `dace:cpu` mimicking the current default behavior.

This change is in anticipation of the `dace:cpu_kfirst` backend that we
already have in the `nasa/milestone2` branch. With that backend, we will
have three backends that can be orchestrated, two of which are
cpu-based. A boolean `on_cpu` is thus not enough anymore to map out all
the options.

Not all backends can be orchestrated. A check is put in place to catch
people to try to orchestrate a non dace-based backend.