api cleanup: remove ndsl/units.py
The functions are all unused (I've checked pyFV3, pySHiELD, and pace)
and should be removed to keep API surface as small as possible. In the
future we envision NDSL to come with much more potent type checking for
units. Until then, if really needed, these functions can be implemented
trivially by dsl users.