be nice with pyfv3 and loading layouts from namelists
loading the layout from a f90nml namelists, we end up with a layout of
type `list[int]`, e.g. [1, 1]. The partitioner expects a `tuple(int,
int)`, e.g. (1, 1) and since this is at the interface betweeen pyfv3 and
ndsl, nobody complains (probably because we don't use mypy correctly in
pyfv3).