Translate tests: raise if test class can't be found (#259)
* raise if translate test can't be found

while it is expected for a test to fail if that happens, this isn't
proper usage of the `xfail` marker. because doing so will just hide that
tests aren't running, e.g. see
https://github.com/NOAA-GFDL/PySHiELD/pull/65.

* allow skip_test to be passed as ctor argument

skip_test is an option of our translate test system. let's be nice an
allow users of the translate test system to use that option as part of
the constructor arguments.

* cleanup exports from ndsl.initialization

---------

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>