refactor: imports cleanup (#264)
* refactor: imports cleanup

This PR suggests three things related to cleaning up imports of NDSL.

1. split grid_sizer and subtile_grid_sizer
2. import MPI directly from mpi4py
3. cleanup direct gt4py imports in tests

The reason to do 1) is that `GridSizer` is used in the `QuantityFactory` as
well as in `SubtileGridSizer`. Both classes live in the same folder and are
exported as public API through the `__init__` file. To avoid a circular
dependency, I've split out `GridSizer` in its own module that can be
loaded upfront.

The reason we previsosly imported `MPI` from `ndsl.comm.mpi` was that at
some point `mpi4py` was an optional dependency and `ndsl.comm.mpi` would
export a stub if `mpi4py` wasn't present. Now that `mpi4py` is a
required dependency, we can just get `MPI` from there.

The reason for 3) is also historical. By now, we make sure to `import .
from dsl` in the top-level `__init__` at the very top. This ensures that
we can properly set the literal precision for gt4py in `ndsl/dsl/__init__.py`.

* fixup

* Undo mpi changes

* Update tests/dsl/test_stencil_wrapper.py

Co-authored-by: Florian Deconinck <deconinck.florian@gmail.com>

---------

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>
Co-authored-by: Florian Deconinck <deconinck.florian@gmail.com>