ci: all the types (or at least most of them) enforced with `mypy` (#258)
* WIP: mypy with strict(er) type checking

* remove a bunch of type: ignore statements

* remove redundant cast operation

* WIP moaaaar types

* moar types

* all the types

* the cherry on top

* fix path / str update after rebase

* fix up: allow None type in local_comm._get_buffer()

* fixup: don't rename arguments ;)

* be nice with pyfv3 and loading layouts from namelists

loading the layout from a f90nml namelists, we end up with a layout of
type `list[int]`, e.g. [1, 1]. The partitioner expects a `tuple(int,
int)`, e.g. (1, 1) and since this is at the interface betweeen pyfv3 and
ndsl, nobody complains (probably because we don't use mypy correctly in
pyfv3).

* readable legacy namelist flag

* more cleanup

* fixing newly added code

* just change "true" -> true

* remove old mypy.ini

---------

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>