Update gt4py to version 1.0.10
GT4Py just dropped version 1.0.10 yesterday. The release includes
everything we need (and two unrelated gt4py.next commits on top). Since
we are on a mission to come back to mainline and be stricter about
releases, we could release NDSL with that version of GT4Py. I think it
would be nice for the baseline numbers. I wouldn't say we always have to
coordinate NDSL and GT4Py release, but sice we have this nice
coincidence, let's use it.