fix: allow any Comm object in ZarrMonitor
This PR is fallout from adding types in PR #257 and #258. The
`ZarrMonitor` provides a `DummyComm` which is instantiated in case no
`Comm` object is given. The type of the `Comm` object in `ZarrMonitor`
was wrongly limited to that `DummyComm`, which only broke when we
attempted to update the submodule in `pace`.