BREAKING CHANGE: change constructor of `QuantityFactory` (#228)
* Breaking change: QuantityFactory from GridSizer and backend name

Change `QuantityFactory` to initialize from a `GridSizer` (as
previously) and a backend name (new). This effectively hides the
previous `numpy` argument, which is effectively an internal allocator
that users shouldn't need to know about. It's basically what
`from_backend()` was doing before (which is now obsolete and was thus
removed).

This is a BREAKING CHANGE and users will need to update their codes if
they instantiated QuantityFactories themselves. For users relying on the
`boilerplate` module, no changes need to happen.

* Keep QuantityFactory.from_backend() with a deprecation warning

* Extended docstings

This is mainly to force a new run of the pyshild workflow now that
pyshield tests are exclusively using `QuantityFactory.from_backend()`
which is compatible with changes proposed in this PR.

* More updates to docstrings

* fixup after rebase

* Unrelated: tests are supposed to return `None`

* fixup: move method back to current place

---------

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>