refactor: directly use gt_storage in QuantityFactory
In the past, `QuantityFactory` would allow not only allocating with
gt4py storage objects, but also directly from `numpy` or `cupy`. This
ability was removed in PR https://github.com/NOAA-GFDL/NDSL/pull/228.
With that removal comes the opportunity to streamline allocation in
`QuantityFactory`, removing the need for a `Allocator` class in the
middle.