refactor: directly use gt_storage in QuantityFactory (#307)
In the past, `QuantityFactory` would allow not only allocating with
gt4py storage objects, but also directly from `numpy` or `cupy`. This
ability was removed in PR https://github.com/NOAA-GFDL/NDSL/pull/228.
With that removal comes the opportunity to streamline allocation in
`QuantityFactory`, removing the need for a `Allocator` class in the
middle.

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>